Backend/requirements.txt
Flask==3.0.2
pymongo[srv]==4.7.2
python-dotenv==1.0.1
flask-cors==4.0.0
